+++ CHANGELOG.rst
@@ -7,6 +7,9 @@ Changelog
 =========
 
 * Support custom ``User`` models.
+* Fixed a bug where a user could end up half logged in if they didn't complete
+  the two-factor login flow. A user's login flow will now be reset. Requires
+  enabled the included middle: ``allauth_2fa.middleware.AllauthTwoFactorMiddleware``.
 * Minor simplifications of code.
 * Minor updates to documentation.